Various compile fixes for samples and faces
Update a number of the samples to accomodate recent changes to materials, faces
and rotations.  Introduce an optional<T> type (probably should rename to
prevent future clashes with std::optional) to allow nullable types, e.g.,
with bounding boxes and faces.  Also clean up some lingering references to
Face4, and remove a number of deprecated code paths specific to Face4.

+++ three/objects/mesh.h
@@ -25,7 +25,7 @@ public:
   int morphTargetBase;
 
   std::vector<int> morphTargetForcedOrder;
-  std::vector<int> morphTargetInfluences;
+  std::vector<float> morphTargetInfluences;
   std::unordered_map<std::string, int> morphTargetDictionary;
 
   int getMorphTargetIndexByName( const std::string& name );
